+++ packages/client/README.md
@@ -9,9 +9,3 @@ yarn run dev
 ```
 
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
-
-## Learn More
-
-To learn more about the project. Checkout `components/docs.md ` directory
-
-You can start editing the page by modifying `pages/index.tsx`. The page auto-updates as you edit the file.
