+++ packages/client/__tests__/components/landingpage/RebatesMiddleware.test.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import "@testing-library/jest-dom";
+import { RebatesMiddleware } from "../../../components";
+import { render } from "@testing-library/react";
+
+/*
+ *  snapshots of the middleware section components
+ */
+describe("Middleware Section ", () => {
+  it("Should render the middleware components ", () => {
+    const home = render(<RebatesMiddleware />);
+    expect(home).toMatchSnapshot();
+  });
+});
