+++ packages/client/components/auth/private.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { NextPage, NextComponentType, NextPageContext } from "next";
-import LandingPage from "../landingpage/landing";
+import { NextPage } from "next";
+import { LandingPage } from "..";
 
 /* HOC for authentication using private id
  * @params PrivateComponent
