+++ packages/client/components/index.tsx
@@ -1,17 +1,20 @@
-export { default as Landingpage } from "./landingpage/landing";
-export { SecureLandingPage } from "./landingpage/landing";
-export { default as Auth } from "./auth/private";
-export { default as Hero } from "./landingpage/firstSection";
-export { default as SecondSection } from "./landingpage/secondSection";
-export { default as ThirdSection } from "./landingpage/thirdSection";
-export { default as FourthSection } from "./landingpage/fourthSection";
-export { default as FifthSection } from "./landingpage/fifthSection";
-export { default as Subscribe } from "./landingpage/subscribe";
-export { default as JsClientSection } from "./landingpage/jsClientSection";
-export { default as Paynav } from "./landingpage/paynav";
-export { default as Footer } from "./landingpage/footer";
+// Reusable exports
 export { default as Image } from "next/image";
+export { default as Link } from "next/link";
 
+// Landing page components
+export { default as LandingPage } from "./landingpage/Default";
+
+export { default as LandingPageLayout } from "./layouts/Default";
+export { UnSecureLandingPage } from "./landingpage/Default";
+export { default as Hero } from "./landingpage/Hero";
+export { default as RebatesMiddleware } from "./landingpage/RebatesMiddleware";
+export { default as BuildPaymasters } from "./landingpage/BuildPaymasters";
+
+// authentication HOC component
+export { default as Auth } from "./auth/private";
+
+// paymasters list page components
 export { default as PaymastersList } from "./activepaymaster/paymastersListPage";
 export { default as Loadmore } from "./activepaymaster/loadmore";
 export { default as ActiveNav } from "./activepaymaster/activeNav";
