+++ packages/client/components/landingpage/partners.js
@@ -1,0 +1,45 @@
+import zkSynk from "../../public/assets/partners/zkSynk.png";
+import StrakNet from "../../public/assets/partners/StrakNet.png";
+import Braavos from "../../public/assets/partners/Braavos.png";
+import Argrent from "../../public/assets/partners/Argrent.png";
+import Chainlink from "../../public/assets/partners/Chainlink.png";
+import OutlierVC from "../../public/assets/partners/Outlier VC.png";
+import DeveloperDao from "../../public/assets/partners/Developer Dao.png";
+
+export const partners = [
+  {
+    id: 1,
+    name: "zkSynk",
+    logo: zkSynk,
+  },
+  {
+    id: 2,
+    name: "StrakNet",
+    logo: StrakNet,
+  },
+  {
+    id: 3,
+    name: "Braavos",
+    logo: Braavos,
+  },
+  {
+    id: 4,
+    name: "Argrent",
+    logo: Argrent,
+  },
+  {
+    id: 5,
+    name: "Chainlink",
+    logo: Chainlink,
+  },
+  {
+    id: 6,
+    name: "Outlier VC",
+    logo: OutlierVC,
+  },
+  {
+    id: 7,
+    name: "Developer Dao",
+    logo: DeveloperDao,
+  },
+];
