+++ packages/client/package.json
@@ -22,10 +22,10 @@
     "eslint-config-next": "13.1.1",
     "jest": "^29.4.1",
     "next": "13.1.1",
+    "next-router-mock": "^0.9.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "4.9.4"
-    
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.3",
@@ -39,6 +39,5 @@
     "postcss": "^8.4.21",
     "prettier": "^2.8.2",
     "tailwindcss": "^3.2.4"
-    
   }
 }
