+++ packages/client/styles/globals.css
@@ -5,8 +5,7 @@
 @tailwind utilities;
 
 @import "./navbar.css";
-@import "./footer.css";
-@import "./dev/main.css";
+@import "./landing/landingpage.css";
 
 * {
   @apply p-0 m-0 box-border outline-0;
@@ -14,22 +13,7 @@
 
 html,
 body {
-  /* max-width: 100vw; */
-  /* overflow-x: hidden; */
-  /* color: theme(fontFamily[color]); */
-  /* font-family: theme(fontFamily[monserrat]), sans-serif; */
-  /* background: theme(colors[backgroundTheme]); */
   @apply bg-paymasters-dark text-white;
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-p {
-  /* font-family: theme(fontFamily[monserrat]), sans-serif; */
 }
 
 a {
@@ -38,8 +22,8 @@ a {
 
 /*************scaling 100 to 125***********************/
 
-@media (min-width: 1100px) and (-webkit-min-device-pixel-ratio: 1.25) {
+/* @media (min-width: 1100px) and (-webkit-min-device-pixel-ratio: 1.25) {
   .scaleto100 {
     @apply px-20;
   }
-}
+} */
