--- packages/client/styles/landing/jsclientsection.module.css
@@ -1,35 +1,0 @@
-/**********************************js client section **********************  */
-
-.js_client_section {
-  @apply my-32 grid grid-cols-[12] grid-rows-4 px-5 gap-6;
-}
-.clients_content {
-  @apply row-start-3 flex justify-center px-3;
-}
-
-.js_client_section .clients_content > div > h5 {
-  @apply text-xl font-medium my-2;
-}
-
-.prog_langs {
-  @apply row-start-4 grid-cols-12;
-}
-
-.prog_langs > h5 {
-  @apply text-xl font-medium;
-}
-.prog_langs .sub_prog_langs {
-  @apply flex flex-wrap justify-start pt-4 gap-1;
-}
-
-@media screen and (max-width: 474px) {
-  .js_client_section {
-    @apply invisible;
-  }
-}
-
-@media (min-width: 475px) and (max-width: 1107px) {
-  .js_client_section {
-    @apply invisible;
-  }
-}
