--- packages/client/styles/landing/secondsection.module.css
@@ -1,84 +1,0 @@
-/* second section components container   */
-
-.second_section {
-  @apply min-h-[65vh] grid px-5 grid-cols-12 gap-5;
-}
-
-.chain_section {
-  @apply col-start-1 col-end-6;
-}
-
-.chain_section > h2 {
-  @apply text-5xl font-semibold;
-}
-
-.chain_image {
-  @apply lg:shrink-0 col-start-8 col-end-12 content-center mb-5 transition-all ease-in duration-300;
-}
-
-.chain_image:hover {
-  @apply scale-110;
-}
-
-.img_blur {
-  @apply col-start-6 col-end-13 h-full bg-[theme(colors[blurColor])]
-    blur-[200px];
-}
-.chain_image:hover {
-  @apply scale-110;
-}
-
-.chain_section > p {
-  @apply text-base font-normal;
-}
-
-/*=================== media queries  ========================================= */
-
-@media screen and (max-width: 474px) {
-  .second_section {
-    @apply grid grid-cols-12 grid-rows-6 my-40 px-3 gap-2 h-screen;
-  }
-
-  .chain_section {
-    @apply col-start-1 col-end-12  row-start-1 row-end-2 mb-8;
-  }
-
-  .chain_section > h2 {
-    @apply text-3xl font-semibold;
-  }
-
-  .chain_image {
-    @apply col-start-1 col-end-9 content-center row-start-6 row-end-7 transition-all ease-in duration-300;
-  }
-  .chain_image img {
-    @apply max-w-[300px];
-  }
-  .img_blur {
-    @apply col-start-1 col-end-12 row-start-5 row-end-7 h-full bg-[theme(colors[blurColor])]
-        blur-[200px];
-  }
-}
-
-@media (min-width: 475px) and (max-width: 1024px) {
-  .second_section {
-    @apply grid grid-cols-12 grid-rows-6 px-3 my-40 gap-2 h-screen;
-  }
-
-  .chain_section {
-    @apply col-start-1 col-end-12  row-start-1 row-end-2 mb-8;
-  }
-
-  .chain_section > h2 {
-    @apply text-3xl font-medium;
-  }
-
-  .chain_image {
-    @apply col-start-2 col-end-12 content-center row-start-4 row-end-5 transition-all ease-in duration-300;
-  }
-  .img_blur {
-    @apply col-start-1 col-end-12 h-full bg-[theme(colors[blurColor])]
-        blur-[300px];
-  }
-}
-
-
