--- packages/client/styles/landing/thirdsection.module.css
@@ -1,186 +1,0 @@
-/******************************* third section components container ************  */
-
-.third_section {
-  @apply h-screen grid grid-cols-12 px-5 grid-rows-6;
-}
-
-.third_section_first_content {
-  @apply col-start-1 col-end-4 text-center row-start-3  row-end-4;
-}
-
-.third_section_second_content {
-  @apply col-start-5 col-end-8 text-center row-start-1 row-end-2;
-}
-
-.third_section_third_content {
-  @apply col-start-9 col-end-13 row-start-3  row-end-4 text-center;
-}
-
-.third_section_first_content span {
-  @apply w-full flex justify-center;
-}
-
-.third_section_second_content span {
-  @apply w-full flex justify-center;
-}
-
-.third_section_third_content span {
-  @apply w-full flex justify-center;
-}
-
-.third_section_img {
-  @apply col-start-5 col-end-8 row-start-4 row-end-5 text-center;
-}
-
-.flower_image_blur {
-  @apply h-[300px] col-start-3 col-end-10 bg-[theme(colors[flowerBlur])] blur-[200px]
-    row-start-2 row-end-6;
-}
-
-.overlap_flower {
-  @apply grid grid-cols-12 grid-rows-6 h-[500px]  col-start-1 col-end-[-1] row-start-5 w-full
-    px-0 border-t-[.8px] border-t-slate-700 bg-backgroundTheme;
-}
-
-.build_paymaster > h2 {
-  @apply text-5xl font-semibold;
-}
-
-.code_snapshot {
-  @apply col-start-3 col-end-8 row-start-3 row-end-5;
-}
-
-.build_blur {
-  @apply col-start-2 col-end-3 row-start-2 row-end-5 bg-[theme(colors[blurColor])]
-    blur-[200px];
-}
-
-.build_paymaster {
-  @apply col-start-7 col-end-12 row-start-3 row-end-5;
-}
-
-/* ===========================  medis queries ==================================================== */
-
-@media screen and (max-width: 474px) {
-  .third_section {
-    @apply h-screen grid grid-cols-12 mt-96 w-full px-2 grid-rows-[repeat(6,_350px)];
-  }
-
-  .third_section_first_content {
-    @apply col-start-1 col-end-12 text-center  gap-12 mb-10 row-start-1;
-  }
-
-  .third_section_second_content {
-    @apply col-start-1 col-end-12 text-center gap-12 row-start-2;
-  }
-
-  .third_section_third_content {
-    @apply col-start-1 col-end-12 text-center gap-12 row-start-3;
-  }
-  .third_section_img {
-    @apply text-center col-start-1 z-0 col-end-12 row-start-4;
-  }
-
-  .flower_image_blur {
-    @apply h-[300px] col-start-2 col-end-12  bg-[theme(colors[flowerBlur])] blur-[200px]
-        row-start-4 row-end-4;
-  }
-
-  .overlap_flower {
-    @apply grid grid-cols-12 grid-rows-[repeat(6,_2fr)] h-[500px] z-10 mt-40 mb-20 col-start-1 col-end-[-1] row-start-4 w-screen
-        px-0 border-t-2 border-t-slate-700 bg-backgroundTheme;
-  }
-
-  .build_paymaster > h2 {
-    @apply text-2xl font-bold;
-  }
-
-  .code_snapshot {
-    @apply col-start-1 col-end-10  row-start-2 row-end-3  mt-20;
-  }
-
-  .code_snapshot img {
-    @apply max-w-none;
-  }
-
-  .build_blur {
-    @apply col-start-2 col-end-5 row-start-2 row-end-3 bg-[theme(colors[blurColor])]
-        blur-[200px];
-  }
-
-  .build_paymaster {
-    @apply col-start-2 col-end-10  row-start-3  row-end-4;
-  }
-}
-
-/* =============================================================================== */
-@media (min-width: 475px) and (max-width: 1024px) {
-  .third_section {
-    @apply h-[70vh] grid grid-cols-12  mx-auto rounded-md grid-rows-[repeat(6,_320px)];
-  }
-
-  .third_section_first_content {
-    @apply col-start-3 col-end-10  text-center px-3 gap-12 mb-10 row-start-1;
-  }
-
-  .third_section_second_content {
-    @apply col-start-3 col-end-10 text-center px-3 gap-12 row-start-2;
-  }
-
-  .third_section_third_content {
-    @apply col-start-3 col-end-10 text-center px-3 gap-12 row-start-3;
-  }
-  .third_section_img {
-    @apply text-center col-start-3 z-0 col-end-12 row-start-4;
-  }
-
-  .flower_image_blur {
-    @apply h-[300px] col-start-2 col-end-12  bg-flowerBlur blur-[200px]
-    row-start-4 row-end-4;
-  }
-
-  .overlap_flower {
-    @apply grid grid-cols-12 grid-rows-[repeat(6,_2fr)] h-[500px] z-10 my-40 rounded-t-lg col-start-1 col-end-[-1] 
-    row-start-4 w-full
-    px-0 border-t-2 border-t-slate-700 bg-backgroundTheme;
-  }
-
-  .build_paymaster > h2 {
-    @apply text-3xl font-semibold;
-  }
-
-  .code_snapshot {
-    @apply col-start-3 col-end-11  row-start-2 row-end-3  mt-20;
-  }
-
-  .code_snapshot img {
-    @apply max-w-none;
-  }
-
-  .build_blur {
-    @apply col-start-2 col-end-5 row-start-2 row-end-3 bg-[theme(colors[blurColor])]
-    blur-[200px];
-  }
-
-  .build_paymaster {
-    @apply col-start-3 text-center  col-end-11 mb-10 row-start-3  row-end-5;
-  }
-}
-
-/*************Large***********************************************************************************/
-
-/******************************************************************************************************/
-
-@media (min-width: 1100px) and (max-width: 1500px) {
-  .third_section_first_content {
-    @apply col-start-1 col-end-4 text-center row-start-2  row-end-3;
-  }
-
-  .third_section_second_content {
-    @apply col-start-5 col-end-9 text-center row-start-1 row-end-2;
-  }
-
-  .third_section_third_content {
-    @apply col-start-9 col-end-13 row-start-2  row-end-3 text-center;
-  }
-}
