+++ packages/client/tailwind.config.js
@@ -9,8 +9,6 @@ module.exports = {
     extend: {
       colors: {
         backgroundTheme: "#11142b",
-        secondaryGlow: "#7d0cc1",
-        jsClientGlow: "#7D0CC1",
         jsClientDark: "rgba(125, 12, 193, 0.3)",
         blurColor: "rgba(125, 12, 193, 0.5)",
         flowerBlur: "rgba(193, 12, 12, 0.5)",
